fill_data_webapp/requirements.txt
fastapi==0.117.1
psycopg2-binary==2.9.10
faker==37.8.0
pydantic-settings==2.10.1
ruff==0.13.1
uvicorn==0.36.0
pytest==8.4.2
pytest-mock==3.15.1
coverage==7.10.7
pre-commit==4.3.0
mypy==1.18.2
types-psycopg2==2.9